Reorganized API view tests
- Each resource now resides in its own file
- Mixins now live in their own file
- Update validate Make target to run quality checks first. (If there are quality issues we can learn of, and fix, them sooner.)

ECOM-3973
